Extracting common client functionality to commons package
